requirement.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
cssselect==1.0.3
Cython==0.28.2
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.2
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.0
django-reversion==2.0.13
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-jwt-proxy==0.1
docutils==0.14
drf-extensions==0.3.1
elasticsearch==6.2.0
elasticsearch-dsl==6.1.0
et-xmlfile==1.0.1
fake-useragent==0.1.10
future==0.16.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.22
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
jwt==0.5.3
Kivy==1.10.0
Kivy-Garden==0.1.4
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
oauthlib==2.0.7
odfpy==1.3.6
openpyxl==2.5.2
parsel==1.4.0
Pillow==5.1.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pycryptodome==3.6.1
PyDispatcher==2.0.5
pygame==1.9.3
Pygments==2.2.0
PyJWT==1.6.1
PyMySQL==0.8.0
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.3
pywin32==223
PyYAML==3.12
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
Scrapy==1.5.0
scrapy-djangoitem==1.1.1
selenium==3.12.0
service-identity==17.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tablib==0.12.1
Twisted==18.4.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
w3lib==1.19.0
wxPython==4.0.0b2
xadmin==2.0.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
yolk==0.4.3
zope.interface==4.5.0
